requirements.txt
Django==4.0
django-redis==4.12.1
redis==3.5.3
orjson==3.6.8
django-ninja==0.17.0
PyJWT==2.0.1
psycopg2==2.9.3
Pillow==9.1.1
oss2==2.5.0
pypinyin==0.41.0